+++ src/shared/role/role.guard.ts
@@ -1,6 +1,6 @@
 import { Injectable, CanActivate, ExecutionContext } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
-import { User } from '../schemas/user.schema';
+import { User } from '../../user/schemas/user.schema';
 
 @Injectable()
 export class RoleGuard implements CanActivate {
